% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dev.R
\name{build_dev}
\alias{build_dev}
\title{Build development materials}
\usage{
build_dev(name)
}
\arguments{
\item{name}{name of lesson}
}
\value{
development folder
}
\description{
Build development materials
}
\examples{
build_dev("foo")
}
