% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_prod.R
\name{build_prod}
\alias{build_prod}
\title{Build Production Materials}
\usage{
build_prod(name)
}
\arguments{
\item{name}{}
}
\value{
production materials in \verb{inst/prod/}
}
\description{
Build Production Materials
}
\examples{
build_prod("test")
}
