% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_prod.R
\name{check_dev_dir}
\alias{check_dev_dir}
\title{Check if development deck exists}
\usage{
check_dev_dir(name)
}
\arguments{
\item{name}{name of folder or lesson}
}
\value{
relative path to slides in \verb{inst/dev/}
}
\description{
Check if development deck exists
}
\examples{
check_dev_dir("test")
}
