% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dev.R
\name{copy_css_styles}
\alias{copy_css_styles}
\title{Copy CSS template to development css folder}
\usage{
copy_css_styles(name)
}
\arguments{
\item{name}{name of folder/lesson}
}
\value{
css style file in inst/dev
}
\description{
Copy CSS template to development css folder
}
\examples{
copy_css_styles("foo")
}
