% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_prod.R
\name{copy_dev_data}
\alias{copy_dev_data}
\title{Copy development data folder to production}
\usage{
copy_dev_data(name)
}
\arguments{
\item{name}{name of development folder}
}
\value{
copied data folder to \code{inst/prod/data}
}
\description{
Copy development data folder to production
}
\examples{
copy_dev_data("test")
}
