% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_prod.R
\name{copy_dev_img}
\alias{copy_dev_img}
\title{Copy development image folder to production}
\usage{
copy_dev_img(name)
}
\arguments{
\item{name}{name of development folder}
}
\value{
copied image folder to \code{inst/prod/img}
}
\description{
Copy development image folder to production
}
\examples{
copy_dev_img("test")
}
