% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_prod.R
\name{copy_dev_slides}
\alias{copy_dev_slides}
\title{Copy development pdf slides to production}
\usage{
copy_dev_slides(name)
}
\arguments{
\item{name}{name of development folder}
}
\value{
copied pdf slides in \code{inst/dev}
}
\description{
Copy development pdf slides to production
}
\examples{
copy_dev_slides("test")
}
