% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_prod.R
\name{copy_dev_worksheets}
\alias{copy_dev_worksheets}
\title{Copy development worksheet folder to production}
\usage{
copy_dev_worksheets(name)
}
\arguments{
\item{name}{name of development folder}
}
\value{
copied worksheet folder to \code{inst/prod/worksheets}
}
\description{
Copy development worksheet folder to production
}
\examples{
copy_dev_worksheets("test")
}
