% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_lesson.R
\name{copy_lesson}
\alias{copy_lesson}
\title{Copy zipped lesson from development into production}
\usage{
copy_lesson(zipped)
}
\arguments{
\item{zipped}{zipped lesson folder}
}
\value{
copied zip folder into production
}
\description{
Copy zipped lesson from development into production
}
