% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_rproj_template.R
\name{copy_rproj_template}
\alias{copy_rproj_template}
\title{Copy .Rproj file to production folder}
\usage{
copy_rproj_template(folder)
}
\arguments{
\item{folder}{name of folder/project}
}
\value{
folder in inst/prod
}
\description{
Copy .Rproj file to production folder
}
\examples{
copy_rproj_template("test")
}
