% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dev.R
\name{create_dev_slide_dir}
\alias{create_dev_slide_dir}
\title{Create development folder in \verb{inst/dev/}}
\usage{
create_dev_slide_dir(name)
}
\arguments{
\item{name}{name of folder}
}
\value{
folder in dev
}
\description{
Create development folder in \verb{inst/dev/}
}
\examples{
create_dev_slide_dir("foo")
}
