% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dev.R
\name{create_dev_slide_file}
\alias{create_dev_slide_file}
\title{Create the development slide .Rmd file}
\usage{
create_dev_slide_file(name)
}
\arguments{
\item{name}{name of development folder/file}
}
\value{
slides.Rmd file
}
\description{
Create the development slide .Rmd file
}
\examples{
create_dev_slide_file("foo")
}
