% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_prod.R
\name{create_pdf_slides}
\alias{create_pdf_slides}
\title{Create development pdf slides}
\usage{
create_pdf_slides(name)
}
\arguments{
\item{name}{name of development folder}
}
\value{
pdf slides in \code{inst/dev}
}
\description{
Create development pdf slides
}
\examples{
create_pdf_slides("test")
}
