% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_prod.R
\name{create_prod_dir}
\alias{create_prod_dir}
\title{Create production folder}
\usage{
create_prod_dir(name)
}
\arguments{
\item{name}{name of development folder to move into production}
}
\value{
folder in \code{inst/prod}
}
\description{
Create production folder
}
\examples{
create_prod_dir("test")
}
