% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_prod.R
\name{create_rproj_file}
\alias{create_rproj_file}
\title{Create .Rproj file in production folder}
\usage{
create_rproj_file(name)
}
\arguments{
\item{name}{name of folder/project}
}
\value{
folder in inst/prod
}
\description{
Create .Rproj file in production folder
}
\examples{
create_rproj_file("test")
}
