% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dev.R
\name{create_slide_folders}
\alias{create_slide_folders}
\title{Create slide development folders}
\usage{
create_slide_folders(name)
}
\arguments{
\item{name}{name of folder/lesson}
}
\value{
css and data folders in inst/dev
}
\description{
Create slide development folders
}
\examples{
create_slide_folders("foo")
}
