% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dev.R
\name{get_slides_rmd}
\alias{get_slides_rmd}
\title{Get \code{slides.Rmd} file from \verb{inst/dev/} folder}
\usage{
get_slides_rmd(name)
}
\arguments{
\item{name}{name of development folder}
}
\value{
slides.Rmd file
}
\description{
Get \code{slides.Rmd} file from \verb{inst/dev/} folder
}
\examples{
get_slides_rmd("foo")
}
