% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_lesson.R
\name{new_lesson}
\alias{new_lesson}
\title{Move lesson from development into production}
\usage{
new_lesson(lesson)
}
\arguments{
\item{lesson}{name of lesson}
}
\value{
copied zip folder into production
}
\description{
Move lesson from development into production
}
\examples{
new_lesson("import")
}
