% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dev.R
\name{remove_dev_dir}
\alias{remove_dev_dir}
\title{Remove development folder}
\usage{
remove_dev_dir(name)
}
\value{
removed folder message
}
\description{
Remove development folder
}
\examples{
remove_prod_dir("foo")
}
