% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_prod.R
\name{remove_prod_dir}
\alias{remove_prod_dir}
\title{Remove production folder}
\usage{
remove_prod_dir(name)
}
\value{
removed folder message
}
\description{
Remove production folder
}
\examples{
remove_prod_dir("test")
}
