% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dev.R
\name{write_slide_header}
\alias{write_slide_header}
\title{Write slide header for slides.Rmd}
\usage{
write_slide_header(name)
}
\arguments{
\item{name}{name of slides and development folder}
}
\value{
slides file with header
}
\description{
Write slide header for slides.Rmd
}
\examples{
write_slide_header("foo")
}
