% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dev.R
\name{write_slide_meta}
\alias{write_slide_meta}
\title{Write meta code chunk for slides.Rmd}
\usage{
write_slide_meta(name)
}
\arguments{
\item{name}{name of slides and development folder}
}
\value{
slides file with meta chunk
}
\description{
Write meta code chunk for slides.Rmd
}
\examples{
write_slide_meta("foo")
}
