% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dev.R
\name{write_slide_setup}
\alias{write_slide_setup}
\title{Write setup code chunk for slides.Rmd}
\usage{
write_slide_setup(name)
}
\arguments{
\item{name}{name of slides and development folder}
}
\value{
slides file with setup chunk
}
\description{
Write setup code chunk for slides.Rmd
}
\examples{
write_slide_setup("foo")
}
