% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dev.R
\name{write_slide_yaml}
\alias{write_slide_yaml}
\title{Write YAML head for slides.Rmd}
\usage{
write_slide_yaml(name)
}
\arguments{
\item{name}{name of slides and development folder}
}
\value{
slides file with YAML
}
\description{
Write YAML head for slides.Rmd
}
\examples{
write_slide_yaml("foo")
}
