% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_lesson.R
\name{zip_lesson}
\alias{zip_lesson}
\title{Zip development lesson}
\usage{
zip_lesson(lesson)
}
\arguments{
\item{lesson}{folder with lesson materials}
}
\value{
compressed folder
}
\description{
Zip development lesson
}
